requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain and MCP Adapters
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-mcp-adapters==0.1.9

# MCP (Model Context Protocol) and Server
mcp>=1.9.2
starlette>=0.27.0
click>=8.0.0

# OpenAI (for LLM)
openai==1.98.0

# Pinecone Vector Database
pinecone>=3.0.0

# Database Connections
psycopg2-binary>=2.9.0
awswrangler[postgres]>=3.0.0

# Environment and Configuration
python-dotenv==1.1.1

# Progress Tracking
tqdm>=4.65.0

# HTTP and Async Support
requests>=2.0.0
aiohttp>=3.8.3
httpx>=0.23.0

# Data Validation and Utilities
pydantic>=2.7.4

# Memory and Storage
sqlalchemy>=2.0.0